Complete Update Motion Estimation
Change of idea: first we compute global motion and object motion in separate way and then we merge them togheter. From testing it result computationa expensive and not so efficient.

The new aproach is to compute motion only in one step: global and local. This result on a more eficient and efficace method.